safari-600.7-branch/LayoutTests/fast/mediastream/RTCPeerConnection-datachannel-expected.txt

Tests RTCDataChannel.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dc = pc.createDataChannel("label", {maxRetransmitTime:0, maxRetransmits:0}); threw exception Error: SyntaxError: DOM Exception 12.
PASS dc = pc.createDataChannel("label1"); did not throw exception.
PASS dc = pc.createDataChannel("label2", {}); did not throw exception.
PASS dc = pc.createDataChannel("label3", {ordered:true}); did not throw exception.
PASS dc = pc.createDataChannel("label3", {ordered:false}); did not throw exception.
PASS dc = pc.createDataChannel("label3", {maxRetransmits:0}); did not throw exception.
PASS dc = pc.createDataChannel("label3", {maxRetransmitTime:0}); did not throw exception.
PASS pc is connected
PASS dc = pc.createDataChannel("label"); did not throw exception.
PASS dc.readyState is 'connecting'
PASS pc_ondatachannel was called
PASS dc_onopen was called
PASS dc.readyState is 'open'
PASS dc.label is 'label'
PASS dc.send('xyzzy'); did not throw exception.
PASS dc_onmessage_string was called
PASS data is 'xyzzy'
PASS dc.send(buffer); did not throw exception.
PASS dc_onmessage_arraybuffer was called
PASS data.byteLength is 2
PASS array[0] is 17
PASS array[1] is 19
PASS dc.send(array); did not throw exception.
PASS dc_onmessage_arraybufferview was called
PASS data.byteLength is 2
PASS array[0] is 11
PASS array[1] is 13
PASS dc_onclose was called
PASS dc.readyState is 'closed'
PASS successfullyParsed is true

TEST COMPLETE
